backend/requirements.txt
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
mysqlclient==1.3.14
pyjwt==1.4.2
